course_project/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
torch>=1.9.0
gpytorch>=1.5.0
scikit-learn>=1.0.0
pyyaml>=5.4.0
tqdm>=4.62.0
properscoring>=0.1
xgboost>=1.5.0
lightgbm>=3.3.0
tensorly>=0.6.0

